+++ src/fix/pointer-focus-children.js
@@ -13,14 +13,14 @@
   It is fixed in IE12 (Win10 IE Tec Preview)
 */
 
-import platform from 'platform';
 import getFocusTarget from '../get/focus-target';
 import decorateContext from '../util/decorate-context';
+import platform from '../util/platform';
 
 let engage;
 let disengage;
 // This fix is only relevant to IE10 (Trident/6) and IE11 (Trident/7)
-const relevantToCurrentBrowser = platform.name === 'IE' && (platform.version.match(/^(10|11)\./));
+const relevantToCurrentBrowser = platform.is.TRIDENT && (platform.is.IE10 || platform.is.IE11);
 // IE10 requires prefix, IE11 does not
 const eventName = typeof document !== 'undefined' && ('onpointerdown' in document ? 'pointerdown' : 'MSPointerDown');