+++ src/query/tabsequence.js
@@ -2,9 +2,9 @@
 // http://www.w3.org/TR/html5/editing.html#sequential-focus-navigation-and-the-tabindex-attribute
 // http://www.w3.org/WAI/PF/aria-practices/#keyboard
 
-import platform from 'platform';
 import queryTabbable from './tabbable';
 import nodeArray from '../util/node-array';
+import platform from '../util/platform';
 import sortArea from './tabsequence.sort-area';
 import sortShadowed from './tabsequence.sort-shadowed';
 import sortTabindex from './tabsequence.sort-tabindex';
@@ -42,7 +42,7 @@ export default function({context, includeContext, strategy} = {}) {
   const _context = nodeArray(context)[0] || document.documentElement;
   let elements = queryTabbable({context: _context, includeContext, strategy});
 
-  if (document.body.createShadowRoot && (platform.name === 'Chrome' || platform.name === 'Chrome Mobile')) {
+  if (document.body.createShadowRoot && platform.is.BLINK) {
     // sort tabindex localized to shadow dom
     // see https://github.com/medialize/ally.js/issues/6
     elements = sortShadowed(elements, _context, sortElements);