+++ src/supports/focus-object-swf.js
@@ -1,8 +1,8 @@
 
-import platform from 'platform';
+import platform from '../util/platform';
 
 // Every Environment except IE9 considers SWF objects focusable
-const result = platform.name !== 'IE' || parseFloat(platform.version) >= 10;
+const result = !platform.is.TRIDENT || !platform.is.IE9;
 
 export default function() {
   return result;